Move propTypes in to the files
