+++ generatedTypes/components/radioButton/RadioGroup.d.ts
@@ -1,5 +1,13 @@
 import React from 'react';
-import { RadioGroupPropTypes, RadioGroupContextPropTypes } from './types';
-export declare const RadioGroupContext: React.Context<RadioGroupContextPropTypes>;
+interface RadioGroupPropTypes {
+    /**
+     * The initial value of the selected radio button
+     */
+    initialValue?: string | boolean;
+    /**
+     * Invoked once when value changes, by selecting one of the radio buttons in the group
+     */
+    onValueChange?: Function;
+}
 declare const _default: React.ComponentType<RadioGroupPropTypes>;
 export default _default;
