+++ src/components/radioButton/RadioButton.tsx
@@ -4,6 +4,7 @@ import {
   StyleSheet,
   Animated,
   Easing,
+  TextStyle,
   StyleProp,
   ImageStyle,
   ViewStyle,
@@ -16,13 +17,68 @@ import View from '../view';
 import Text from '../text';
 const Image = require('../image').default; // TODO: change this once Image is written in typescript
 import asRadioGroupChild from './asRadioGroupChild';
-import {RadioButtonPropTypes, RadioButtonState, RadioGroupChildPropTypes, RadioGroupContextPropTypes} from './types';
+import {RadioGroupContextPropTypes} from './RadioGroupContext';
 
 const DEFAULT_SIZE = 24;
 const DEFAULT_COLOR = Colors.blue30;
 
+interface RadioButtonPropTypes {
+  /**
+   * The identifier value of the radio button. must be different than other RadioButtons in the same group
+   */
+  value?: string | boolean;
+  /**
+   * When using RadioButton without a RadioGroup, use this prop to toggle selection
+   */
+  selected?: boolean;
+  /**
+   * Invoked when pressing the button
+   */
+  onPress?: Function;
+  /**
+   * Whether the radio button should be disabled
+   */
+  disabled?: boolean;
+  /**
+   * The color of the radio button
+   */
+  color?: string;
+  /**
+   * The size of the radio button, affect both width & height
+   */
+  size?: number;
+  /**
+   * The radio button border radius
+   */
+  borderRadius?: number;
+  /**
+   * A label for the radio button description
+   */
+  label?: string;
+  /**
+   * Label style
+   */
+  labelStyle?: TextStyle;
+  /**
+   * Icon image source
+   */
+  iconSource?: object | number;
+  /**
+   * Icon image style
+   */
+  iconStyle?: ImageStyle;
+  /**
+   * Should the icon be on the right side of the label
+   */
+  iconOnRight?: boolean;
+}
+
+interface RadioButtonState {
+  opacityAnimationValue: Animated.Value;
+  scaleAnimationValue: Animated.Value;
+}
+
 type Props = RadioButtonPropTypes &
-  RadioGroupChildPropTypes &
   RadioGroupContextPropTypes &
   BaseComponentInjectedProps &
   ForwardRefInjectedProps &
