+++ src/components/radioButton/asRadioGroupChild.tsx
@@ -1,9 +1,21 @@
 import React, {Component} from 'react';
-import {RadioGroupContext} from './RadioGroup';
-import {RadioGroupChildPropTypes} from './types';
+import RadioGroupContext from './RadioGroupContext';
+
+interface RadioGroupChildPropTypes {
+  /**
+   * The identifier value of the radio button. must be different than other RadioButtons in the same group
+   */
+  value?: string | boolean;
+  /**
+   * When using RadioButton without a RadioGroup, use this prop to toggle selection
+   */
+  selected?: boolean;
+}
+
+type PropTypes = RadioGroupChildPropTypes;
 
 export default function asRadioGroupChild(WrappedComponent: React.ComponentType<any>) {
-  class RadioGroupChild extends Component<RadioGroupChildPropTypes> {
+  class RadioGroupChild extends Component<PropTypes> {
     render() {
       const {value: buttonValue, selected} = this.props;
       return (
